flask_gorgumu/requirements.txt
flask==2.2.3
flask-login==0.6.2
werkzeug==2.2.3
cassandra-driver==3.25.0
ecdsa==0.18.0
qrcode==7.4.2
pillow==9.5.0
zmq==0.0.0
python-dotenv==1.0.0
pycryptodome==3.17.0
Markdown==3.4.3
based58==0.1.1
python-dateutil==2.8.2
requests==2.28.2 